[FLINK-16660] Make rpc timeout and retry period configurable for application mode
In Application Mode, the user code is executed on the cluster but goes
through rpc (to the dispatcher gateway). Currently, the closest available
option is the WebOptions.TIMEOUT, but this does not cover both the timeout
and the retry period, and it is made for another reason. To this end, we
add two options specifically for this.