[FLINK-20001] Don't use setAllVerticesInSameSlotSharingGroupByDefault in StreamGraphGenerator
I think the default of having all vertices in the same slot sharing
group should be good for both BATCH and STREAMING right now. We can
reconsider actually setting this flag in the future.

Background information: this is a special setting that was introduced
for the Table API Blink runner/planner, which does special things with
slot sharing (or lack thereof) and is more aware of memory requirements
and such. For general DataStream/DataSet programs changing the setting
doesn't make sense.